Code refactoring
Pie chart names transformation into array done in .jsp now
ExpenseKey class moved to separate file.
