+++ src/main/java/com/es/jointexpensetracker/service/ExpenseService.java
@@ -2,6 +2,7 @@ package com.es.jointexpensetracker.service;
 
 import com.es.jointexpensetracker.exception.DataNotFoundException;
 import com.es.jointexpensetracker.model.Expense;
+import com.es.jointexpensetracker.model.ExpenseKey;
 
 import java.util.List;
 
@@ -11,7 +12,7 @@ public interface ExpenseService {
 
     List<Expense> getExpensesByGroup(String expenseGroup);
 
-    Expense loadExpenseByKey(Expense.ExpenseKey key) throws DataNotFoundException;
+    Expense loadExpenseByKey(ExpenseKey key) throws DataNotFoundException;
 
     Expense addExpense(String expenseGroup);
